+++ src/os/posix/src/os-impl-timebase.c
@@ -277,7 +277,7 @@ int32 OS_Posix_TimeBaseAPI_Impl_Init(void)
         * Pre-calculate the clock tick to microsecond conversion factor.
         * This is used by OS_Tick2Micros(), OS_Milli2Ticks(), etc.
         */
-       OS_SharedGlobalVars.TicksPerSecond = sysconf(_SC_CLK_TCK);
+       OS_SharedGlobalVars.TicksPerSecond = (int32)sysconf(_SC_CLK_TCK);
        if (OS_SharedGlobalVars.TicksPerSecond <= 0)
        {
           OS_DEBUG("Error: Unable to determine OS ticks per second: %s\n",strerror(errno));
@@ -420,7 +420,7 @@ int32 OS_TimeBaseCreate_Impl(uint32 timer_id)
              * The output is irrelevant here; the objective is to just ensure
              * that the signal is not already pending.
              */
-            i = sysconf( _SC_SIGQUEUE_MAX);
+            i = (int)sysconf( _SC_SIGQUEUE_MAX);
             do
             {
                 ts.tv_sec = 0;
@@ -507,8 +507,8 @@ int32 OS_TimeBaseSet_Impl(uint32 timer_id, int32 start_time, int32 interval_time
         ** Convert from Microseconds to timespec structures
         */
         memset(&timeout, 0, sizeof(timeout));
-        OS_UsecToTimespec(start_time, &timeout.it_value);
-        OS_UsecToTimespec(interval_time, &timeout.it_interval);
+        OS_UsecToTimespec((uint32)start_time, &timeout.it_value);
+        OS_UsecToTimespec((uint32)interval_time, &timeout.it_interval);
 
         /*
         ** Program the real timer
